+++ front-end/src/components/product-table-row.component.js
@@ -19,10 +19,12 @@ export default class ProductTableRow extends Component {
 
     async deleteOneProduct() {
         try {
-            const res = await ProductService.deleteProduct(this.state.currentUser.id, this.props.product._id);
-            // Refresh to refresh category table
-            //TODO: trigger a rerender of the table only instead of the entire page
-            window.location.reload();
+            if (window.confirm("Are you sure you wish to delete this?")) {
+                const res = await ProductService.deleteProduct(this.state.currentUser.id, this.props.product._id);
+                // Refresh to refresh category table
+                //TODO: trigger a rerender of the table only instead of the entire page
+                window.location.reload();
+            }
         } catch (err) {
             alert("Error deleting product");
         }
